advance io and random number: create separate implementation for c and c++ print: fix function argument (not checked in standard) analog io: remove reference to gpio pwm and mux (define) and set to dummy value digital io: fix typo and use hexadecimal instead of binary for constant value external interrupt: set dummy branch for compilation only wifi server: add constructor and destructor serial: remove uart and fix return value in void function stream: fix typo for type/class String variables: add not a pin definition

+++ arduino/cores/Print.cpp
@@ -83,8 +83,8 @@ size_t Print::println(void)
   return write("\r\n");
 }
 
-size_t Print::print(double val, uint8_t format){
-    return printFloat(val, format);
+size_t Print::print(double val, int format){
+    return printFloat(val, (uint8_t) format);
 }
 
 size_t Print::println(unsigned char val){
